(Fix) move google analytics to layout
