--- app/head.tsx
@@ -1,21 +1,0 @@
-import Script from 'next/script';
-
-export default function Head() {
-  return (
-    <>
-      {/* Google Analytics */}
-      <Script async src="https://www.googletagmanager.com/gtag/js?id=G-9WM6MNGPP8" />
-      <Script
-        id="google-analytics"
-        dangerouslySetInnerHTML={{
-          __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', 'G-9WM6MNGPP8');
-          `,
-        }}
-      />
-    </>
-  );
-}
