+++ app/layout.tsx
@@ -2,6 +2,7 @@ import './globals.css';
 
 import { AppLayout } from '@/src/app/ui/AppLayout';
 import { Providers } from '@/src/app/ui/Providers';
+import GoogleAnalytics from '@/src/shared/lib/GoogleAnalytics';
 
 import React from 'react';
 import type { Metadata } from 'next';
@@ -22,6 +23,9 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en">
       <body className="antialiased">
+        {process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS && (
+          <GoogleAnalytics googleAnalyticsId={process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS} />
+        )}
         <Script
           src={`//dapi.kakao.com/v2/maps/sdk.js?appkey=${kakaoKey}&autoload=false&libraries=clusterer,drawing`}
           strategy="beforeInteractive"
